fix(Praditor): 修复参数文件处理逻辑并更新测试文件
修复参数文件处理中的异常捕获逻辑，将默认参数文件名从params.txt改为params_vad.txt
更新测试音频的参数文件和文本网格文件，删除不再使用的测试文件
在核心代码中添加调试打印语句

+++ Praditor.py
@@ -347,13 +347,16 @@ class MainWindow(QMainWindow):
         #     with open("params.txt", 'w') as txt_file:
         #         txt_file.write(f"{self.MySliders.getParams()}")
         # else:
-        try:
-            with open("params.txt", "r") as txt_file:
-                self.MySliders.resetParams(eval(txt_file.read()))
-        except KeyError or OSError:
+        if not os.path.exists("params.txt"):
             with open("params.txt", 'w') as txt_file:
                 txt_file.write(f"{self.MySliders.getParams()}")
-
+        else:  # 存在default mode
+            try:
+                with open("params.txt", "r") as txt_file:
+                    self.MySliders.resetParams(eval(txt_file.read()))
+            except KeyError:
+                with open("params.txt", 'w') as txt_file:
+                    txt_file.write(f"{self.MySliders.getParams()}")
